libraries/pcbgenerator/requirements.txt
freetype-py
svg.path
fonttools
pytest
pytest-timeout
cairosvg
pillow
shapely
